support multiple examples
