+++ src/renderer/constants/routes.ts
@@ -1,4 +1,7 @@
 export default {
   ROOT: '/',
-  MODELLING_UI_POC: '/modelling-ui-poc',
+  MODELLING_UI_POC: {
+    ROOT: '/modelling-ui-poc',
+    SIMPLE_CUBE: '/modelling-ui-poc/simple-cube',
+  },
 };
