+++ src/renderer/pages/modelling-ui-poc/index.ts
@@ -1,1 +1,4 @@
+/* eslint-disable import/prefer-default-export */
+
 export { default } from './modelling-ui-poc';
+export { default as SimpleCube } from './simple-cube';
