+++ src/renderer/pages/modelling-ui-poc/simple-cube/index.ts
@@ -1,0 +1,1 @@
+export { default } from './simple-cube';
